+++ src/libcsg/imcio.cc
@@ -41,7 +41,7 @@ void imcio_write_dS(const string &file, ub::vector<double> &r, ub::vector<double
         throw runtime_error(string("error, cannot open file ") + file);
 
     if(list == NULL) {
-        for (int i = 0; i < dS.size(); ++i) {
+        for (size_t i = 0; i < dS.size(); ++i) {
             out_dS << r[i] << " " << dS[i] << endl;
         }
     }
@@ -95,7 +95,7 @@ void imcio_write_index(const string &file, vector<string> &names, vector<RangePa
     if (!out_idx)
         throw runtime_error(string("error, cannot open file ") + file);
     
-    for (int i = 0; i < names.size(); ++i)
+    for (size_t i = 0; i < names.size(); ++i)
         out_idx << names[i] << " " << ranges[i] << endl;
     
     out_idx.close();
@@ -125,7 +125,7 @@ void imcio_read_matrix(const string &filename, ub::matrix<double> &gmc)
     if(!in)
         throw runtime_error(string("error, cannot open file ") + filename);
 
-    int line_count =0;
+    size_t line_count =0;
     string line;
     // read till the first data line
     while(getline(in, line)) {
@@ -147,7 +147,7 @@ void imcio_read_matrix(const string &filename, ub::matrix<double> &gmc)
         if(gmc.size1()!=tokens.size())
             throw runtime_error(string("error loading ")
                     + filename + ": size mismatchm, number of columns differ");
-        for(int i=0; i<tokens.size(); ++i)
+        for(size_t i=0; i<tokens.size(); ++i)
             gmc(line_count,i) = boost::lexical_cast<double>(tokens[i]);
         ++line_count;
     }
